added dummy insert method

+++ BusChecker/CoreData/BusStopMO.swift
@@ -11,7 +11,7 @@ import CoreData
 
 class BusStopMO : NSManagedObject {
     
-    class func CreateBusStopMO(stopCode : Int16,bookmarkName: String,busNumber: String, creation: Date, in context: NSManagedObjectContext) -> BusStopMO
+    class func CreateBusStopMO(stopCode : Int64,bookmarkName: String,busNumber: String, creation: Date, in context: NSManagedObjectContext) -> BusStopMO
     {
     
         let busstop = BusStopMO(context: context)
